templates/gpg: disable protection
Without this option, gpg wants the user to enter a passphrase which
fails after switching to {{ aptly-user }} with su.
The issue is described here: https://github.com/MISP/MISP/issues/3702

Nevertheless, the password (if any) should be provided non-interactive.
If a password is needed, '%no-protection' should be replaced by
'Passphrase: <string>'